Parse TypeScript type aliases

+++ example/rust/src/lib.rs
@@ -2,7 +2,7 @@ use crate::types::User;
 
 pub mod types;
 
-pub fn create_user(id: u64, name: String, company_email: String) -> User {
+pub fn create_user(id: u32, name: String, company_email: String) -> User {
     User {
         id,
         name,
